servers/VerifyServer/pyproject.toml
[project]
name = "verifyserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0",
]
